Use StreamReader for both hybi and draft75.
